+++ src/Gate.IO.Api/Futures/Responses/GateFuturesBalanceChange.cs
@@ -33,7 +33,7 @@ public record GateFuturesBalanceChange
     /// <summary>
     /// Changing Type：
     /// </summary>
-    [JsonProperty("type")]
+    [JsonProperty("type"), JsonConverter(typeof(MapConverter))]
     public GateFuturesBalanceChangeType Type { get; set; }
 
     /// <summary>